Improve output on multi-bridge systems
This commit fixes the problem of losing flood/fast-leave ports from
bridges not named 'br0'.  It also fixes the long-standing issue of not
showing any router ports at all.

To maintain backwards compatibility with Westermo WeOS, we keep the
old code as "compat".  It relies mostly on custom, unoffial patches
to the kernel, and will probably be removed completely in the next
major version.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>